Add term mapping lookup ...
Due to the terms query not ordering results based off the terms
provided, we have to store a pointer to the result associated with
each identifier so that the user provided CURIE's appear in the expected
order. This *should* be lightweight given it shouldn't copy the data and
is only a pointer to the corresponding data, but the lookup table is
quite large given the number of equivalent identifiers is roughly double
the amount of provided CURIE's from what I've seen during testing.

Overall the handler is setup now but a little messy. The performance
seems appropriate for local, but we shall have to see have the
performance with more network latency