Add get_normalized_nodes endpoint to nodenorm API (#299)
* Update the .gitignore for biothings_hub folders

* Add the get_normalized_nodes handler for nodenorm API

* Add nodenorm handler tests

* Add the new endpoint to the config_web API_LIST

* Add the biolink model toolkit to requirements.txt

* Add malformed CURIE handling to the handler response

* Update the source fields to include all identifiers

* Add term mapping lookup ...

Due to the terms query not ordering results based off the terms
provided, we have to store a pointer to the result associated with
each identifier so that the user provided CURIE's appear in the expected
order. This *should* be lightweight given it shouldn't copy the data and
is only a pointer to the corresponding data, but the lookup table is
quite large given the number of equivalent identifiers is roughly double
the amount of provided CURIE's from what I've seen during testing.

* Update the parser to include conflation information

* Initial attempt to add all conflation backend logic