Nodenorm handle curie duplication (#325)
* Transition to using a different multiprocessing approach ...

Initially attempted a thread pool for connecting to MongoDB, 
but it appeared that the amount of work was still CPU limited so 
wasn't seeing significant gains from parallelization. Switched to 
a ProcessPool to ensure we bypass GIL

* Add identifier sqlite3 database generation

* Add a post_update function for parsing the CURIE duplication identifiers

* Add curie deduplication handling ...

All workers track their identifiers and return them from 
the process pool to be aggregated into a sqlite3 database

We then search for all duplicate CURIE entries and handle
them on a individual basis depending on 1 of 4 cases that
apply based off the conflict between document insertions

* Add an identifier.i mongodb index for FIND speedup

* Add a sqlite3 index for SELECT speedup

* Separate all uploading logic to the worker module

* Ensure we have copies of all arguments for the upload_process callback

* Ensure the upload_process job returns an integer

* Add a retry mechanism for the CURIE duplication connection ...

Sometimes we'd have connection issues during the `job_manager`
which appeared to stall the threads attempting a connection.

This should mitigate that as the heartbeat only lasts about 10 seconds

* Update README

* Remove the parse module (incorporated into worker.py)

* Default to cpu count for dumper workers